requirements.txt
psycopg2-binary==2.9.9
uvicorn==0.30.6
fastapi==0.115.0
gunicorn==23.0.0
SQLAlchemy[async]==2.0.35
python-dotenv==1.0.1
pydantic-settings==2.5.2
itsdangerous==2.2.0
requests==2.32.3
uuid-utils==0.9.0
boto3==1.35.22
APScheduler==3.10.4
python-jose[cryptography]==3.3.0
asyncpg==0.29.0
redis[hiredis]==5.0.8
aiohttp==3.10.5
kombu==5.4.1
celery==5.4.0
asgiref==3.8.1
pandas==2.2.2
elasticsearch==8.15.1
